refactor: rename namespaces
